+++ src/interlegis/portalmodelo/policy/tests/test_discussion_settings.py
@@ -13,22 +13,17 @@ class DiscussionTestCase(unittest.TestCase):
 
     def setUp(self):
         self.portal = self.layer['portal']
-        self.workflow = self.portal['portal_workflow']
 
     def test_configuration(self):
         settings = getUtility(IRegistry).forInterface(IDiscussionSettings)
         self.assertTrue(settings.anonymous_comments)
         self.assertTrue(settings.anonymous_email_enabled)
-        self.assertTrue(settings.moderation_enabled)
+        self.assertFalse(settings.moderation_enabled)
         self.assertEqual(settings.captcha, u'captcha')
         self.assertTrue(settings.globally_enabled)
         self.assertTrue(settings.show_commenter_image)
-        self.assertEqual(settings.text_transform, u'text/plain')
-
-    def test_workflow(self):
-        self.assertIn('comment_review_workflow', self.portal.portal_workflow)
-        chain = self.workflow.getChainForPortalType('Discussion Item')
-        self.assertEqual(chain, ('comment_review_workflow',))
+        self.assertTrue(settings.user_notification_enabled)
+        self.assertEqual(settings.text_transform, u'text/x-web-intelligent')
 
     def test_news_items_have_discussion_enabled(self):
         news_item = self.portal['portal_types']['News Item']